+++ datastores/executions.go
@@ -1,7 +1,7 @@
 package datastores
 
 type Execution struct {
-	UserID int    `db:"user_id"`
+	UserID int64  `db:"user_id"`
 	Token  string `db:"token" json:"-"`
 }
