+++ datastores/tokens.go
@@ -7,6 +7,6 @@ import (
 )
 
 type Tokenstore interface {
-	SaveToken(userID int, token *oauth2.Token) error
-	Schedule(userID int, date time.Time) error
+	SaveToken(userID int64, token *oauth2.Token) error
+	Schedule(userID int64, date time.Time) error
 }
